+++ web/src/data/local/index.ts
@@ -1,0 +1,1 @@
+export { LocalVolumeService } from './LocalVolumeService';
