+++ web/src/data/translator/index.ts
@@ -2,5 +2,4 @@ export { translate } from './api';
 export { Translator } from './translator';
 export type { TranslatorId } from './translator';
 
-export { PersonalVolumesManager } from './db/personal';
-export { createSegIndexedDbCache } from './db/segment_cache';
+export { CachedSegRepository } from './SegCacheRepository';
