+++ web/src/pages/novel/components/WebTranslate.vue
@@ -11,7 +11,7 @@ import {
   useGptWorkspaceStore,
   useSakuraWorkspaceStore,
 } from '@/data/stores/workspace';
-import { PersonalVolumesManager } from '@/data/translator/db/personal';
+import { LocalVolumeService } from '@/data/local';
 
 import TranslateOptions from './TranslateOptions.vue';
 
@@ -85,9 +85,9 @@ const files = computed(() => {
 const importToWorkspace = async () => {
   const blob = await ky.get(files.value.jp.url).blob();
   const file = new File([blob], files.value.jp.filename);
-  await PersonalVolumesManager.saveVolume(file)
+  await LocalVolumeService.createVolume(file)
     .then(() =>
-      PersonalVolumesManager.updateGlossary(file.name, toRaw(props.glossary))
+      LocalVolumeService.updateGlossary(file.name, toRaw(props.glossary))
     )
     .then(() => message.success('导入成功'))
     .catch((error) => message.error(`导入失败:${error}`));
