+++ web/src/pages/workspace/SakuraWorkspace.vue
@@ -10,7 +10,7 @@ import { ref } from 'vue';
 import { notice } from '@/components/NoticeBoard.vue';
 import { useSettingStore } from '@/data/stores/setting';
 import { TranslateJob, useSakuraWorkspaceStore } from '@/data/stores/workspace';
-import { createSegIndexedDbCache } from '@/data/translator';
+import { CachedSegRepository } from '@/data/translator';
 import { useIsWideScreen } from '@/data/util';
 import SoundAllTaskCompleted from '@/sound/all_task_completed.mp3';
 
@@ -80,8 +80,7 @@ const onProgressUpdated = (
 };
 
 const clearCache = async () => {
-  const cache = await createSegIndexedDbCache('sakura-seg-cache');
-  await cache.clear();
+  await CachedSegRepository.clear('sakura-seg-cache');
   message.success('缓存清除成功');
 };
